Tests: isolate DBF schema rewrite memo regression
