+++ tests/test_dbf_table_support.h
@@ -44,5 +44,6 @@ void test_pack_memo_preserves_payloads_that_decode_empty();
 void test_pack_memo_preserves_binary_picture_payloads();
 void test_pack_memo_fails_when_referenced_payload_cannot_be_recovered();
 void test_additive_memo_replace_preserves_raw_payload_and_fails_closed();
+void test_schema_rewrites_preserve_raw_memo_and_unaffected_field_bytes();
 
 }  // namespace copperfin::test_dbf_table
